Move all the turmoil tests into a Turmoil package.
Also s/GlobalEventsDealer/GlobalEventDealer
