+++ tests/turmoil/Turmoil.spec.ts
@@ -1,30 +1,30 @@
 import {expect} from 'chai';
-import {Player} from '../src/Player';
-import {PartyName} from '../src/turmoil/parties/PartyName';
-import {Game} from '../src/Game';
-import {MarsFirst} from '../src/turmoil/parties/MarsFirst';
-import {Phase} from '../src/Phase';
-import {OrOptions} from '../src/inputs/OrOptions';
-import {SelectSpace} from '../src/inputs/SelectSpace';
-import {SpaceBonus} from '../src/SpaceBonus';
-import {Turmoil} from '../src/turmoil/Turmoil';
-import {TestingUtils, setCustomGameOptions, TestPlayers} from './TestingUtils';
-import {Reds} from '../src/turmoil/parties/Reds';
-import {ReleaseOfInertGases} from '../src/cards/base/ReleaseOfInertGases';
-import {JovianEmbassy} from '../src/cards/promo/JovianEmbassy';
-import {IceAsteroid} from '../src/cards/base/IceAsteroid';
-import {ProtectedValley} from '../src/cards/base/ProtectedValley';
-import {MagneticFieldGeneratorsPromo} from '../src/cards/promo/MagneticFieldGeneratorsPromo';
-import {Resources} from '../src/Resources';
-import {NitrogenFromTitan} from '../src/cards/colonies/NitrogenFromTitan';
-import {SpaceStation} from '../src/cards/base/SpaceStation';
-import {EarthCatapult} from '../src/cards/base/EarthCatapult';
-import {QuantumExtractor} from '../src/cards/base/QuantumExtractor';
-import * as constants from '../src/constants';
-import {SerializedTurmoil} from '../src/turmoil/SerializedTurmoil';
-import {PoliticalAgendas} from '../src/turmoil/PoliticalAgendas';
-import {IParty} from '../src/turmoil/parties/IParty';
-import {GreeneryStandardProject} from '../src/cards/base/standardProjects/GreeneryStandardProject';
+import {Player} from '../../src/Player';
+import {PartyName} from '../../src/turmoil/parties/PartyName';
+import {Game} from '../../src/Game';
+import {MarsFirst} from '../../src/turmoil/parties/MarsFirst';
+import {Phase} from '../../src/Phase';
+import {OrOptions} from '../../src/inputs/OrOptions';
+import {SelectSpace} from '../../src/inputs/SelectSpace';
+import {SpaceBonus} from '../../src/SpaceBonus';
+import {Turmoil} from '../../src/turmoil/Turmoil';
+import {TestingUtils, setCustomGameOptions, TestPlayers} from '../TestingUtils';
+import {Reds} from '../../src/turmoil/parties/Reds';
+import {ReleaseOfInertGases} from '../../src/cards/base/ReleaseOfInertGases';
+import {JovianEmbassy} from '../../src/cards/promo/JovianEmbassy';
+import {IceAsteroid} from '../../src/cards/base/IceAsteroid';
+import {ProtectedValley} from '../../src/cards/base/ProtectedValley';
+import {MagneticFieldGeneratorsPromo} from '../../src/cards/promo/MagneticFieldGeneratorsPromo';
+import {Resources} from '../../src/Resources';
+import {NitrogenFromTitan} from '../../src/cards/colonies/NitrogenFromTitan';
+import {SpaceStation} from '../../src/cards/base/SpaceStation';
+import {EarthCatapult} from '../../src/cards/base/EarthCatapult';
+import {QuantumExtractor} from '../../src/cards/base/QuantumExtractor';
+import * as constants from '../../src/constants';
+import {SerializedTurmoil} from '../../src/turmoil/SerializedTurmoil';
+import {PoliticalAgendas} from '../../src/turmoil/PoliticalAgendas';
+import {IParty} from '../../src/turmoil/parties/IParty';
+import {GreeneryStandardProject} from '../../src/cards/base/standardProjects/GreeneryStandardProject';
 
 describe('Turmoil', function() {
   let player : Player; let player2 : Player; let game : Game; let turmoil: Turmoil;
