+++ tests/turmoil/globalEvents/GlobalEventDealer.spec.ts
@@ -1,17 +1,17 @@
 import {expect} from 'chai';
-import {Game} from '../src/Game';
-import {GlobalDustStorm} from '../src/turmoil/globalEvents/GlobalDustStorm';
-import {GlobalEventDealer, getGlobalEventByName} from '../src/turmoil/globalEvents/GlobalEventDealer';
-import {GlobalEventName} from '../src/turmoil/globalEvents/GlobalEventName';
-import {IGlobalEvent} from '../src/turmoil/globalEvents/IGlobalEvent';
-import {ScientificCommunity} from '../src/turmoil/globalEvents/ScientificCommunity';
-import {SerializedGlobalEventDealer} from '../src/turmoil/globalEvents/SerializedGlobalEventDealer';
-import {SponsoredProjects} from '../src/turmoil/globalEvents/SponsoredProjects';
-import {SuccessfulOrganisms} from '../src/turmoil/globalEvents/SuccessfulOrganisms';
-import {WarOnEarth} from '../src/turmoil/globalEvents/WarOnEarth';
-import {setCustomGameOptions, TestPlayers} from './TestingUtils';
+import {Game} from '../../../src/Game';
+import {GlobalDustStorm} from '../../../src/turmoil/globalEvents/GlobalDustStorm';
+import {GlobalEventDealer, getGlobalEventByName} from '../../../src/turmoil/globalEvents/GlobalEventDealer';
+import {GlobalEventName} from '../../../src/turmoil/globalEvents/GlobalEventName';
+import {IGlobalEvent} from '../../../src/turmoil/globalEvents/IGlobalEvent';
+import {ScientificCommunity} from '../../../src/turmoil/globalEvents/ScientificCommunity';
+import {SerializedGlobalEventDealer} from '../../../src/turmoil/globalEvents/SerializedGlobalEventDealer';
+import {SponsoredProjects} from '../../../src/turmoil/globalEvents/SponsoredProjects';
+import {SuccessfulOrganisms} from '../../../src/turmoil/globalEvents/SuccessfulOrganisms';
+import {WarOnEarth} from '../../../src/turmoil/globalEvents/WarOnEarth';
+import {setCustomGameOptions, TestPlayers} from '../../TestingUtils';
 
-describe('GlobalEventsDealer', () => {
+describe('GlobalEventDealer', () => {
   it('serialize/deserialize - empty', () => {
     const dealer = new GlobalEventDealer([], []);
