+++ tests/turmoil/parties/Scientists.spec.ts
@@ -1,12 +1,12 @@
 import {expect} from 'chai';
-import {Player} from '../../src/Player';
-import {Game} from '../../src/Game';
-import {Turmoil} from '../../src/turmoil/Turmoil';
-import {TestingUtils, setCustomGameOptions, TestPlayers} from '../TestingUtils';
-import {Scientists, SCIENTISTS_BONUS_1, SCIENTISTS_BONUS_2, SCIENTISTS_POLICY_1, SCIENTISTS_POLICY_4} from '../../src/turmoil/parties/Scientists';
-import {SearchForLife} from '../../src/cards/base/SearchForLife';
-import {Research} from '../../src/cards/base/Research';
-import {GeneRepair} from '../../src/cards/base/GeneRepair';
+import {Player} from '../../../src/Player';
+import {Game} from '../../../src/Game';
+import {Turmoil} from '../../../src/turmoil/Turmoil';
+import {TestingUtils, setCustomGameOptions, TestPlayers} from '../../TestingUtils';
+import {Scientists, SCIENTISTS_BONUS_1, SCIENTISTS_BONUS_2, SCIENTISTS_POLICY_1, SCIENTISTS_POLICY_4} from '../../../src/turmoil/parties/Scientists';
+import {SearchForLife} from '../../../src/cards/base/SearchForLife';
+import {Research} from '../../../src/cards/base/Research';
+import {GeneRepair} from '../../../src/cards/base/GeneRepair';
 
 describe('Scientists', function() {
   let player : Player; let game : Game; let turmoil: Turmoil; let scientists: Scientists;
