+++ tests/turmoil/parties/Unity.spec.ts
@@ -1,14 +1,14 @@
 import {expect} from 'chai';
-import {Player} from '../../src/Player';
-import {Game} from '../../src/Game';
-import {Turmoil} from '../../src/turmoil/Turmoil';
-import {TestingUtils, setCustomGameOptions, TestPlayers} from '../TestingUtils';
-import {Unity, UNITY_BONUS_1, UNITY_BONUS_2, UNITY_POLICY_2, UNITY_POLICY_3} from '../../src/turmoil/parties/Unity';
-import {SisterPlanetSupport} from '../../src/cards/venusNext/SisterPlanetSupport';
-import {VestaShipyard} from '../../src/cards/base/VestaShipyard';
-import {LocalShading} from '../../src/cards/venusNext/LocalShading';
-import {OrOptions} from '../../src/inputs/OrOptions';
-import {Tags} from '../../src/cards/Tags';
+import {Player} from '../../../src/Player';
+import {Game} from '../../../src/Game';
+import {Turmoil} from '../../../src/turmoil/Turmoil';
+import {TestingUtils, setCustomGameOptions, TestPlayers} from '../../TestingUtils';
+import {Unity, UNITY_BONUS_1, UNITY_BONUS_2, UNITY_POLICY_2, UNITY_POLICY_3} from '../../../src/turmoil/parties/Unity';
+import {SisterPlanetSupport} from '../../../src/cards/venusNext/SisterPlanetSupport';
+import {VestaShipyard} from '../../../src/cards/base/VestaShipyard';
+import {LocalShading} from '../../../src/cards/venusNext/LocalShading';
+import {OrOptions} from '../../../src/inputs/OrOptions';
+import {Tags} from '../../../src/cards/Tags';
 
 describe('Unity', function() {
   let player : Player; let game : Game; let turmoil: Turmoil; let unity: Unity;
